Create necessary files and structures for the project
This commit includes the following changes:
- Create include and src folder
- Move several input functions to an input file
- Create hotel function signatures
