+++ include/input.h
@@ -1,0 +1,11 @@
+#include<string.h> 
+#include<stdlib.h>
+#include<stdio.h>
+#include<ctype.h>
+
+char input_char( char letter);
+int input_dec_number(int number);
+double input_time(double time);
+
+// Andrew's part
+double input_fee();
