+++ src/hotel.c
@@ -1,0 +1,6 @@
+#include "../include/hotel.h"
+
+double get_input_fee()
+{
+  
+}
